refactor evalArg into two separate functions
We only care about the result of evalArg in evalChainNode, where the
formal type is always nil. In that case validateType always returns
argType. So we split evalArg into two funcs, one where the formal is
nil and we care about the result, and one where it isn't and we don't.