add break and continue nodes
Add cases for the new (as of 1.18) break and continue nodes.

These have no direct type-checking implications, so we just ignore them.
However, they do affect control flow, so they can result in
dead code which could affect type checking. For example, in this code:

	{{$v := 1}}
	{{range .}}
		{{break}}
		{{$v = ""}}
	{{end}}
	{{$v.I}}

the assignment to $v in the loop is dead, so we know that $v is a
number after the loop and can report an error. We don't do that,
because it would complicate the checker for minimal benefit (people
rarely write dead code).

Fixes #3.